// models/SignUp.Model.ts
export interface SignupData {
  title: "Mr" | "Mrs";
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  day: string;
  month: string;
  year: string;
  company: string;
  address1: string;
  address2: string;
  country: string;
  state: string;
  city: string;
  zipcode: string;
  mobileNumber: string;
  newsletter: boolean;
  specialOffers: boolean;
}